% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_components.R
\name{widget_gauge}
\alias{widget_gauge}
\title{Colour-coded category gauge}
\usage{
widget_gauge(widget_data)
}
\arguments{
\item{widget_data}{A \code{quarto_widget_data} object from \code{\link{widget_data}}.}
}
\value{
An \code{htmltools::tag} (\code{<div id="<id>-gauge">}).
}
\description{
Places the \code{<div>} for the horizontal category gauge.
}
\examples{
\dontrun{
wd <- widget_data(df, id = "demo")
wd
widget_gauge(wd)
}

}
