% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_components.R
\name{widget_header}
\alias{widget_header}
\title{Detail header}
\usage{
widget_header(widget_data)
}
\arguments{
\item{widget_data}{A \code{quarto_widget_data} object from \code{\link{widget_data}}.}
}
\value{
An \code{htmltools::tag} (\code{<div id="<id>-detail-header">}).
}
\description{
Places the \code{<div>} for the detail panel heading.
}
\examples{
\dontrun{
wd <- widget_data(df, id = "demo")
wd
widget_header(wd)
}

}
