% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst_dashboard.R
\name{widget_table}
\alias{widget_table}
\title{Detail indicator table}
\usage{
widget_table(widget_data)
}
\arguments{
\item{widget_data}{A \code{quarto_widget_data} object returned by
\code{\link{widget_data}}.}
}
\value{
An \code{htmltools::tag} (\code{<div>}) with the id
\code{"<id>-table-output"}. Content is managed by the JavaScript runtime.
}
\description{
Places the container \code{<div>} for the detail indicator table. When the
user hovers over a sunburst segment the table is populated with the
indicators belonging to the hovered node for the currently selected entity.
The table is rendered by DataTables and includes the score and any
reference values defined in \code{comparison_cols}.
}
\examples{
\dontrun{
wd <- widget_data(df, hierarchy, id = "demo")
wd
widget_header(wd)
widget_table(wd)
widget_plot(wd)
}

}
\seealso{
\code{\link{widget_data}}, \code{\link{widget_header}},
\code{\link{widget_plot}}, \code{\link{widget_layout}}
}
